DXP-1672 Rename basic data loaders
