+++ src/catalog/Model/Indexer/DataLoader/CategoryDataLoader.php
@@ -1,14 +1,14 @@
 <?php declare(strict_types=1);
 
-namespace StreamX\ConnectorCatalog\Model\Indexer\Action;
+namespace StreamX\ConnectorCatalog\Model\Indexer\DataLoader;
 
 use Exception;
 use Magento\Framework\Exception\NoSuchEntityException;
 use StreamX\ConnectorCatalog\Model\ResourceModel\Category as ResourceModel;
-use StreamX\ConnectorCore\Api\BaseAction;
+use StreamX\ConnectorCore\Api\BasicDataLoader;
 use Traversable;
 
-class CategoryAction implements BaseAction {
+class CategoryDataLoader implements BasicDataLoader {
 
     private ResourceModel $resourceModel;
