+++ src/core/Api/BasicDataLoader.php
@@ -4,11 +4,10 @@ namespace StreamX\ConnectorCore\Api;
 
 use Traversable;
 
-interface BaseAction
+interface BasicDataLoader
 {
 
     /**
-     * Loads basic data for the entities by IDs.
      * @param array $entityIds if empty - loads data for all available IDs for the entity type
      */
     public function loadData(int $storeId, array $entityIds): Traversable;
